requirements.txt
bs4~=0.0.2
beautifulsoup4~=4.12.3
loguru~=0.7.2
pyfiglet~=1.0.2
PyMySQL~=1.1.1
selenium~=4.25.0
redis~=5.0.8
tqdm~=4.66.5
setuptools~=58.0.4
python-dotenv~=1.0.1
requests~=2.32.3
python-dotenv~=1.0.1
fastapi~=0.115.6
pytest~=8.3.3